Users are now added to the database when they create an account
Also now Answer is correctly populated with answered and unanswered
questions for this user. Meant to make this two separate commits but had
to go back and amend the commit that did the users correctly and it
added all of my changes for answer too!

(From initial user database commit) Though the only information collected during account creation is display name, which asks for first and last name but doesn't enforce (my code handles the case where they only give one name) - so we should create/redirect to an activity where they fill out the rest of their info