Fix hopefully the last authentication bug.
AuthStateListener was being triggered multiple times before user signed
in, causing multiple sign-in activities to start in a sort of nested
way, so when the user finished one they were sent to the next one even
though they just signed in. This introduces a boolean that is set to
true if the sign-in activity has already started so it only starts once.
Then for robustness, the boolean is set back to false if anything but
success happens from that sign-in activity, and upon sign-out.