AskSelect and Answer work with new schema for friends in the database
New schema is that for user/friends, instead of having a list of friends
that would have fields with their ID and their name (which would have to
be updated when they updated their profile), it just has a list where
the keys are the userIDs of the friends and the values are "True".

Also added a field to user part of the database called fullName which is a
concatenation of firstName+" "+lastName, which is updated at the creation
of the user's account but NOT YET UPDATED upon profile change, since others
are working on that code right now.